refactor: split lua script engine collaborators
Extract cache, module loading, and MoonSharp reflection helpers from LuaScriptEngineService while preserving the IScriptEngineService contract.

Keep bootstrap, plugin loading, meta generation, and file watching in the facade so runtime behavior stays unchanged during the refactor.

Refs #174
